requirements.txt
aio-pika
sqlalchemy
asyncpg
httpx
twilio
sendgrid
python-dotenv
